move specs into corresponding files
